Added and Renamed Sounds
Basically added the full BGM that Andy and I felt fit pretty well. I'm
not sure how the sound files work with the code, but you can clip parts
of it and loop it or you can loop the whole thing, it doesn't really
matter at this point does it?